Anti theft system for your bike
The system sleeps for 5min intervals and wakes up only to read SMS ( or get pressed by a button to wake up). 
when the system is awake it reads an SMS.
if the SMS is not a key word (in this case ON OR OFF) - the system goes back to sleep.
if the SMS is ON - the system goes to ON MODE
ON MODE is the state where the accelerometer checks every 10 sec if the bike have been moved and goes back to sleep.
once the accelerometer decides the bike has been move it sends the user a message stating that the bike are being stolen, in addition it sends the bikes location using google maps.
the system sends this information every 10 sec untill the user sends an "OFF" message. 
when the system receives OFF message - it goes back to the first state (5 min sleep intervals to check sms).